+++ paymentsheet/src/main/java/com/stripe/android/link/ui/wallet/WalletViewModel.kt
@@ -95,6 +95,7 @@ internal class WalletViewModel(
             paymentSelectionHint = paymentSelectionHint,
             signupToggleEnabled = configuration.linkSignUpOptInFeatureEnabled,
             billingDetailsCollectionConfiguration = configuration.billingDetailsCollectionConfiguration,
+            linkBrand = configuration.linkBrand,
         )
     )
 
@@ -124,6 +125,7 @@ internal class WalletViewModel(
     val uiState: StateFlow<WalletUiState> = _uiState.asStateFlow()
 
     val allowLogOut: Boolean = configuration.allowLogOut
+    val linkBrand = configuration.linkBrand
 
     val expiryDateController = SimpleTextFieldController(
         textFieldConfig = DateConfig()