+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/ui/WalletButtonsInteractor.kt
@@ -80,8 +80,8 @@ internal interface WalletButtonsInteractor {
         @Stable
         data class Link(
             val state: LinkButtonState,
+            val linkBrand: LinkBrand,
             val theme: LinkButtonTheme = LinkButtonTheme.DEFAULT,
-            val linkBrand: LinkBrand = LinkBrand.Link,
         ) : WalletButton {
             override val walletType = WalletType.Link
 
@@ -293,7 +293,9 @@ internal class DefaultWalletButtonsInteractor constructor(
                 WalletButton.Link(
                     state = LinkButtonState.Default,
                     theme = arguments.configuration.walletButtons?.buttonThemes?.link
-                        ?: LinkButtonTheme.DEFAULT
+                        ?: LinkButtonTheme.DEFAULT,
+                    // Defaulting to Link because this code path isn't accessible
+                    linkBrand = LinkBrand.Link,
                 ),
                 arguments
             )